enclave: Rename msg_channel to secret_channel
