+++ standalone/pruntime/enclave/src/secret_channel.rs
